requirements-dev.txt
-r requirements.txt
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-httpx>=0.30.0
respx>=0.21.0
